Fixed Performance Issue
Removed Horrible Caching Idea for an LRUCache and an expiration date.
LRUCache set to 4 mb
The cache will invalidate after 24 hours

Pulled out the the Webservice and ImageLoader into separate module for
fun.

Created a combined dependencies in the project gradle file so that I
can have a common area to get my dependencies from removing the chance
of duplicating and mismatching libs. Then compiled the libs into the
correct modules.