Added Empty and Loading View
Added an Empty and Loading view.
Added Observerable that monitors the state of the current view and
tells the user what is going on in within the app at that time.

Added Custom search hints for each user and on the main screen instead
of just Search…

Added connection utils to check if the network is available. if it is
not don’t even try to get new data and just display the data that is
currently there.